# docs/api/openapi.yaml
components:
  schemas:
...
    AdminGrantResultDto:
      additionalProperties: false
      description: Admin manual grant result (subscription or entitlement grant).
      properties:
        code:
          nullable: true
          type: string
        entitlementId:
          format: uuid
          nullable: true
          type: string
        id:
          format: uuid
          type: string
        planId:
          format: uuid
          nullable: true
          type: string
        status:
          nullable: true
          type: string
        userId:
          format: uuid
          type: string
      required:
        - id
        - userId
      type: object
    AdminNotificationResultDto:
      additionalProperties: false
      description: Admin operational notification creation result.
      properties:
        id:
          format: uuid
          type: string
        type:
          type: string
        userId:
          format: uuid
          type: string
      required:
        - id
        - userId
        - type
      type: object
    ApiMeta:
      additionalProperties: true
      description: >-
        List metadata (docs/07_API.md §19). The server normalizes page/limit (max server-enforced limit: 100) and
        computes total/totalPages/hasNext. Clients must never derive totals themselves.
      properties:
        hasNext:
          type: boolean
        limit:
          maximum: 100
          minimum: 1
          type: integer
        page:
          minimum: 1
          type: integer
        total:
          minimum: 0
          type: integer
        totalPages:
          minimum: 0
          type: integer
      type: object
...
    ErrorCode:
      enum:
        - AUTH_REQUIRED
        - AUTH_INVALID
        - FORBIDDEN
        - NOT_FOUND
        - VALIDATION_ERROR
        - CONFLICT
        - RATE_LIMITED
        - SUBSCRIPTION_REQUIRED
        - ENTITLEMENT_REQUIRED
        - INVALID_STATE_TRANSITION
        - UPLOAD_REJECTED
        - INTERNAL_ERROR
      type: string
    ErrorResponse:
      description: Canonical error envelope (docs/07_API.md §3).
      properties:
        error:
          properties:
            code:
              $ref: '#/components/schemas/ErrorCode'
            fields:
              additionalProperties:
                type: string
              description: Field-level validation details when applicable.
              type: object
            message:
              type: string
          required:
            - code
            - message
          type: object
      required:
        - error
      type: object
...
    PaymentMethodConfigDto:
      additionalProperties: false
      description: Admin-managed manual payment destination (method + account + display name).
      properties:
        accountIdentifier:
          type: string
        displayName:
          type: string
        isEnabled:
          type: boolean
        method:
          enum:
            - instapay
            - vodafone_cash
          type: string
      required:
        - method
        - accountIdentifier
        - displayName
        - isEnabled
      type: object
...
    PaymentSubmissionDto:
      additionalProperties: false
      description: Manual payment submission (pending → approved | rejected by ADMIN only).
      properties:
        createdAt:
          format: date-time
          type: string
        id:
          format: uuid
          type: string
        method:
          enum:
            - instapay
            - vodafone_cash
          type: string
        planId:
          format: uuid
          type: string
        proofStorageKey:
          nullable: true
          type: string
        reviewedAt:
          format: date-time
          nullable: true
          type: string
        status:
          enum:
            - pending
            - approved
            - rejected
          type: string
        subscriptionId:
          format: uuid
          nullable: true
          type: string
        transferReference:
          type: string
        updatedAt:
          format: date-time
          type: string
        userId:
          format: uuid
          type: string
      required:
        - id
        - userId
        - planId
        - subscriptionId
        - method
        - transferReference
        - proofStorageKey
        - status
        - reviewedAt
        - createdAt
        - updatedAt
      type: object
...
  securitySchemes:
    bearer:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  description: >-
    Al-Khabir backend contract. Generated from the NestJS controllers (ADR-0003). The canonical success envelope is `{
    data, meta? }`; errors use `{ error: { code, message, fields? } }` with the codes documented in docs/07_API.md §21.
  title: Al-Khabir API
  version: 0.1.0
openapi: 3.0.3
paths:
...
  /api/v1/admin/entitlements/grant:
    post:
      operationId: AdminGrantsController_grantEntitlement
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: false
              properties:
                entitlement_id:
                  type: string
                user_id:
                  type: string
              required:
                - user_id
                - entitlement_id
              type: object
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/AdminGrantResultDto'
                  meta:
                    $ref: '#/components/schemas/ApiMeta'
                required:
                  - data
                type: object
          description: Entitlement granted manually (audited).
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Admin authority required.
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: User or entitlement not found.
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Entitlement already granted / inactive.
      security:
        - bearer: []
      tags:
        - admin
...
  /api/v1/admin/notifications:
    post:
      operationId: AdminNotificationsController_notify
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: false
              properties:
                body_ar:
                  maxLength: 2000
                  minLength: 1
                  type: string
                data_json: {}
                title_ar:
                  maxLength: 255
                  minLength: 1
                  type: string
                type:
                  maxLength: 64
                  minLength: 1
                  type: string
                user_id:
                  type: string
              required:
                - user_id
                - type
                - title_ar
                - body_ar
              type: object
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/AdminNotificationResultDto'
                  meta:
                    $ref: '#/components/schemas/ApiMeta'
                required:
                  - data
                type: object
          description: Operational notification created (audited).
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Admin authority required.
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Recipient user not found.
      security:
        - bearer: []
      tags:
        - admin
  /api/v1/admin/payments/config:
    get:
      operationId: AdminPaymentsController_getConfig
      parameters: []
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/PaymentMethodConfigDto'
                  meta:
                    $ref: '#/components/schemas/ApiMeta'
                required:
                  - data
                type: object
          description: All configured manual payment destinations.
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Validation failed (canonical error envelope).
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Admin authority required.
      security:
        - bearer: []
      tags:
        - admin-payments
  /api/v1/admin/payments/config/{method}:
    put:
      operationId: AdminPaymentsController_upsertConfig
      parameters:
        - in: path
          name: method
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: false
              properties:
                account_identifier:
                  maxLength: 255
                  minLength: 1
                  type: string
                display_name:
                  maxLength: 255
                  minLength: 1
                  type: string
                is_enabled:
                  type: boolean
              required:
                - account_identifier
                - display_name
                - is_enabled
              type: object
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/PaymentMethodConfigDto'
                  meta:
                    $ref: '#/components/schemas/ApiMeta'
                required:
                  - data
                type: object
          description: Payment destination upserted (audited).
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Validation failed (canonical error envelope).
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Admin authority required.
      security:
        - bearer: []
      tags:
        - admin-payments
  /api/v1/admin/payments/submissions:
    get:
      operationId: AdminPaymentsController_listSubmissions
      parameters:
        - in: query
          name: page
          required: false
          schema:
            minimum: 1
            type: integer
        - in: query
          name: limit
          required: false
          schema:
            maximum: 100
            minimum: 1
            type: integer
        - in: query
          name: status
          required: false
          schema:
            enum:
              - pending
              - approved
              - rejected
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/PaymentSubmissionDto'
                  meta:
                    $ref: '#/components/schemas/ApiMeta'
                required:
                  - data
                type: object
          description: Payment submissions (oldest first for the review queue).
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Validation failed (canonical error envelope).
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Admin authority required.
      security:
        - bearer: []
      tags:
        - admin-payments
  /api/v1/admin/payments/submissions/{id}:
    get:
      operationId: AdminPaymentsController_detail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/PaymentSubmissionDto'
                  meta:
                    $ref: '#/components/schemas/ApiMeta'
                required:
                  - data
                type: object
          description: Submission detail (incl. typed proof reference).
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Validation failed (canonical error envelope).
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Admin authority required.
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Payment submission not found.
      security:
        - bearer: []
      tags:
        - admin-payments
  /api/v1/admin/payments/submissions/{id}/approve:
    post:
      operationId: AdminPaymentsController_approve
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/PaymentApprovalResultDto'
                  meta:
                    $ref: '#/components/schemas/ApiMeta'
                required:
                  - data
                type: object
          description: 'Approved: ACTIVE subscription + audit + notification (one transaction).'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Validation failed (canonical error envelope).
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Admin authority required.
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Payment submission not found.
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Already reviewed / plan unavailable (canonical envelope).
      security:
        - bearer: []
      tags:
        - admin-payments
  /api/v1/admin/payments/submissions/{id}/reject:
    post:
      operationId: AdminPaymentsController_reject
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/PaymentRejectResultDto'
                  meta:
                    $ref: '#/components/schemas/ApiMeta'
                required:
                  - data
                type: object
          description: 'Rejected: no activation + audit + notification (one transaction).'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Validation failed (canonical error envelope).
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Admin authority required.
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Payment submission not found.
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Already reviewed (canonical envelope).
      security:
        - bearer: []
      tags:
        - admin-payments
  /api/v1/admin/subscriptions/grant:
    post:
      operationId: AdminGrantsController_grantSubscription
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: false
              properties:
                plan_id:
                  type: string
                user_id:
                  type: string
              required:
                - user_id
                - plan_id
              type: object
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/AdminGrantResultDto'
                  meta:
                    $ref: '#/components/schemas/ApiMeta'
                required:
                  - data
                type: object
          description: ACTIVE subscription granted manually (audited).
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Admin authority required.
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: User or plan not found.
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Active-subscription conflict / plan inactive / role mismatch.
      security:
        - bearer: []
      tags:
        - admin
...
  /api/v1/payments/config:
    get:
      operationId: SubscriptionsController_paymentConfig
      parameters: []
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/PaymentMethodConfigDto'
                  meta:
                    $ref: '#/components/schemas/ApiMeta'
                required:
                  - data
                type: object
          description: Currently enabled manual payment destinations.
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Validation failed (canonical error envelope).
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Missing/malformed token or role mismatch.
      security:
        - bearer: []
      tags:
        - subscriptions
  /api/v1/payments/submissions:
    get:
      operationId: SubscriptionsController_listOwnSubmissions
      parameters:
        - in: query
          name: page
          required: false
          schema:
            minimum: 1
            type: integer
        - in: query
          name: limit
          required: false
          schema:
            maximum: 100
            minimum: 1
            type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/PaymentSubmissionDto'
                  meta:
                    $ref: '#/components/schemas/ApiMeta'
                required:
                  - data
                type: object
          description: Own payment submissions (paginated).
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Validation failed (canonical error envelope).
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Missing/malformed token or role mismatch.
      security:
        - bearer: []
      tags:
        - subscriptions
...